+++ package.json
@@ -6,6 +6,7 @@
     "scripts": {
         "test": "jest --coverage",
         "test-watch": "jest --watch",
+        "tsc": "tsc",
         "eslint": "eslint .",
         "codecov": "codecov"
     },
@@ -31,6 +32,8 @@
     },
     "devDependencies": {
         "@react-native-community/eslint-config": "^3.0.1",
+        "@types/react": "^17.0.44",
+        "@types/react-native": "^0.67.6",
         "babel-eslint": "^10.1.0",
         "babel-jest": "^27.2.5",
         "codecov": "^3.7.2",
@@ -41,7 +44,8 @@
         "prettier": "^2.4.1",
         "react": "^17.0.2",
         "react-native": "^0.66.1",
-        "react-test-renderer": "^17.0.2"
+        "react-test-renderer": "^17.0.2",
+        "typescript": "^4.6.4"
     },
     "files": [
         "ios",
